requirements.txt
torch
opencv-python-headless
rasterio
numpy
shapely
fiona
transformers
scikit-learn
pillow
torchvision
matplotlib
psutil
scikit-image
tkinter